# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Ynov API for YnovImmo"
  version: "1.0.0"
  title: "Ynov API pour l'application YnovImmo"
host: "localhost:8080"
basePath: "/api/v1"
schemes:
- "http"
paths:
  /chat:
    get:
      tags:
      - "chat"
      summary: "get all chat by pagination"
      description: ""
      operationId: "chatGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where\
          \ id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "limit returning object count"
        required: false
        type: "integer"
        default: 2
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ChatPagination"
      x-swagger-router-controller: "Chat"
    post:
      tags:
      - "chat"
      summary: "create chat"
      description: "create chat"
      operationId: "chatPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create chat"
        required: true
        schema:
          $ref: "#/definitions/Chat"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Chat"
    patch:
      tags:
      - "chat"
      summary: "update chat"
      description: "update chat"
      operationId: "chatPATCH"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create chat"
        required: true
        schema:
          $ref: "#/definitions/Chat"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Chat"
  /chat/{Id}:
    get:
      tags:
      - "chat"
      summary: "get a chat by ID"
      description: "get a chat by ID"
      operationId: "chatIdGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of chat to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Chat"
      x-swagger-router-controller: "Chat"
    delete:
      tags:
      - "chat"
      summary: "Destroy a chat by ID"
      description: "delete a chat by ID"
      operationId: "chatIdDELETE"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of chat to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Chat"
  /chat-comment:
    get:
      tags:
      - "chat-comment"
      summary: "get all chat-comment by pagination"
      description: ""
      operationId: "chat_commentGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where\
          \ id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "limit returning object count"
        required: false
        type: "integer"
        default: 2
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ChatCommentPagination"
      x-swagger-router-controller: "ChatComment"
    post:
      tags:
      - "chat-comment"
      summary: "create chat-comment"
      description: "create chat-comment"
      operationId: "chat_commentPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create chat-comment"
        required: true
        schema:
          $ref: "#/definitions/ChatComment"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "ChatComment"
    patch:
      tags:
      - "chat-comment"
      summary: "update chat-comment"
      description: "update chat-comment"
      operationId: "chat_commentPATCH"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create chat-comment"
        required: true
        schema:
          $ref: "#/definitions/ChatComment"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "ChatComment"
  /chat-comment/{Id}:
    get:
      tags:
      - "chat-comment"
      summary: "get a chat-comment by ID"
      description: "get a chat-comment by ID"
      operationId: "chat_commentIdGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of chat-comment to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ChatComment"
      x-swagger-router-controller: "ChatComment"
    delete:
      tags:
      - "chat-comment"
      summary: "Destroy a chat-comment by ID"
      description: "delete a chat-comment by ID"
      operationId: "chat_commentIdDELETE"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of chat-comment to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "ChatComment"
  /city:
    get:
      tags:
      - "city"
      summary: "get all city by pagination"
      description: ""
      operationId: "cityGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where\
          \ id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "limit returning object count"
        required: false
        type: "integer"
        default: 2
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CityPagination"
      x-swagger-router-controller: "City"
    post:
      tags:
      - "city"
      summary: "create city"
      description: "create city"
      operationId: "cityPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create city"
        required: true
        schema:
          $ref: "#/definitions/City"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "City"
    patch:
      tags:
      - "city"
      summary: "update city"
      description: "update city"
      operationId: "cityPATCH"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create city"
        required: true
        schema:
          $ref: "#/definitions/City"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "City"
  /city/{Id}:
    get:
      tags:
      - "city"
      summary: "get a city by ID"
      description: "get a city by ID"
      operationId: "cityIdGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of city to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/City"
      x-swagger-router-controller: "City"
    delete:
      tags:
      - "city"
      summary: "Destroy a city by ID"
      description: "delete a city by ID"
      operationId: "cityIdDELETE"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of city to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "City"
  /cities-district:
    get:
      tags:
      - "cities-district"
      summary: "get all cities-district by pagination"
      description: ""
      operationId: "cities_districtGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where\
          \ id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "limit returning object count"
        required: false
        type: "integer"
        default: 2
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CitiesDistrictPagination"
      x-swagger-router-controller: "CitiesDistrict"
    post:
      tags:
      - "cities-district"
      summary: "create cities-district"
      description: "create cities-district"
      operationId: "cities_districtPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create cities-district"
        required: true
        schema:
          $ref: "#/definitions/CitiesDistrict"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "CitiesDistrict"
    patch:
      tags:
      - "cities-district"
      summary: "update cities-district"
      description: "update cities-district"
      operationId: "cities_districtPATCH"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create cities-district"
        required: true
        schema:
          $ref: "#/definitions/CitiesDistrict"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "CitiesDistrict"
  /cities-district/{Id}:
    get:
      tags:
      - "cities-district"
      summary: "get a cities-district by ID"
      description: "get a cities-district by ID"
      operationId: "cities_districtIdGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of cities-district to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/CitiesDistrict"
      x-swagger-router-controller: "CitiesDistrict"
    delete:
      tags:
      - "cities-district"
      summary: "Destroy a cities-district by ID"
      description: "delete a cities-district by ID"
      operationId: "cities_districtIdDELETE"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of cities-district to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "CitiesDistrict"
...
definitions:
  Chat:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Id de la conversation"
      id_user1:
        type: "integer"
        format: "int64"
        description: "Id utilisateur 1"
      id_user2:
        type: "integer"
        format: "int64"
        description: "Id utilisateur 2"
    example:
      id: 5
      id_user2: 7
      id_user1: 2
  ChatPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
        description: "json repose code"
      total:
        type: "integer"
        format: "int32"
        description: "total numbers"
      offset:
        type: "integer"
        format: "int32"
        description: "offset"
      limit:
        type: "integer"
        format: "int32"
        description: "limit"
      list:
        type: "array"
        items:
          $ref: "#/definitions/Chat"
    example:
      total: 6
      code: 0
      offset: 1
      limit: 5
      list:
      - id: 5
        id_user2: 7
        id_user1: 2
      - id: 5
        id_user2: 7
        id_user1: 2
  ChatComment:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Id du commentaire"
      id_chat:
        type: "integer"
        format: "int64"
        description: "Id de la conversation"
      id_user:
        type: "integer"
        format: "int64"
        description: "Id de l'utilisateur qui a écrit le commentaire"
      comment:
        type: "string"
        format: "string"
        description: "Commentaire"
      date:
        type: "string"
        format: "date-time"
        description: "Date du commentaire"
    example:
      id_chat: 2
      date: "2000-01-23T04:56:07.000+00:00"
      comment: "comment"
      id: 5
      id_user: 7
  ChatCommentPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
        description: "json repose code"
      total:
        type: "integer"
        format: "int32"
        description: "total numbers"
      offset:
        type: "integer"
        format: "int32"
        description: "offset"
      limit:
        type: "integer"
        format: "int32"
        description: "limit"
      list:
        type: "array"
        items:
          $ref: "#/definitions/ChatComment"
    example:
      total: 6
      code: 0
      offset: 1
      limit: 5
      list:
      - id_chat: 2
        date: "2000-01-23T04:56:07.000+00:00"
        comment: "comment"
        id: 5
        id_user: 7
      - id_chat: 2
        date: "2000-01-23T04:56:07.000+00:00"
        comment: "comment"
        id: 5
        id_user: 7
  City:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Id de la ville"
      name:
        type: "string"
        format: "string"
        description: "Nom de la ville"
      zip_code:
        type: "string"
        format: "string"
        description: "Code postal de la ville"
    example:
      name: "name"
      id: 5
      zip_code: "zip_code"
  CityPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
        description: "json repose code"
      total:
        type: "integer"
        format: "int32"
        description: "total numbers"
      offset:
        type: "integer"
        format: "int32"
        description: "offset"
      limit:
        type: "integer"
        format: "int32"
        description: "limit"
      list:
        type: "array"
        items:
          $ref: "#/definitions/City"
    example:
      total: 6
      code: 0
      offset: 1
      limit: 5
      list:
      - name: "name"
        id: 5
        zip_code: "zip_code"
      - name: "name"
        id: 5
        zip_code: "zip_code"
  CitiesDistrict:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Id du quartier"
      id_city:
        type: "integer"
        format: "int64"
        description: "Id de la ville"
      name:
        type: "string"
        format: "string"
        description: "Nom du quartier"
    example:
      id_city: 2
      name: "name"
      id: 5
  CitiesDistrictPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
        description: "json repose code"
      total:
        type: "integer"
        format: "int32"
        description: "total numbers"
      offset:
        type: "integer"
        format: "int32"
        description: "offset"
      limit:
        type: "integer"
        format: "int32"
        description: "limit"
      list:
        type: "array"
        items:
          $ref: "#/definitions/CitiesDistrict"
    example:
      total: 6
      code: 0
      offset: 1
      limit: 5
      list:
      - id_city: 2
        name: "name"
        id: 5
      - id_city: 2
        name: "name"
        id: 5
...
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      msg:
        type: "string"
    example:
      msg: "msg"
      code: 0
...